Model módosítások, Article és Advertising osztályok most már leszármazottjai a PageReservation osztálynak
