+++ Maestro/Model/Advertising.swift
@@ -1,0 +1,13 @@
+//
+//  Advertising.swift
+//  Maestro
+//
+//  Created by Sashalmi Imre on 2023. 03. 16..
+//
+
+import Foundation
+import RealmSwift
+
+class Advertising: PageReservation {
+    
+}
